reactingEulerFoam: Corrected blending in tutorials
Sub-model blending should be set such that the sum of all the blending
coefficients equals one. If there are three models specified for a phase
pair (e.g., (air in water), (water in air) and (air and water)), then
the sum-to-one constraint is guaranteed by the blending functions.
Frequently, however, the symmetric model ((air and water) in this
example) is omitted. In that case, the blending coefficients should be
selected so that the sum of just the two non-symmetric coefficients
equal one.

In the case of linear blending, this means setting the minimum partially
continuous alpha to one-minus the fully continuous value of the opposite
phase. For example:

   blending
   {
       default
       {
           type            linear;
           minFullyContinuousAlpha.air 0.7;
           minPartlyContinuousAlpha.air 0.3;
           minFullyContinuousAlpha.water 0.7;
           minPartlyContinuousAlpha.water 0.3;
       }
   }

The reactingTwoPhaseEulerFoam and reactingMultiPhaseEulerFoam tutorials
have been modified to adhere to this principle.