rigidBodyDynamics: Added some 0-DoF joints
The new "rigid" joint permits no motion at all, "function" specifies the
position of the joint as a function of the position of it's parent, and
"functionDot" specifies the position of the joint as a function of the
velocity of it's parent. Note that the functions are applied uniformly
to each component of the parent joint's position/motion. Example
specifications are shown below.

    joint
    {
        type    rigid;
    }

    joint
    {
        type    function;
        function table ((-1 0) (0 1) (1 0));
    }

    joint
    {
        type    functionDot;
        function table ((-1 0) (0 1) (1 0));
    }

This work was supported by Caitlin Worden Hodge, at Zyba