displacementLinearMotionMotionSolver: New simple linear motion solver
Mesh motion solver simple linear expansion and contraction of a mesh
region defined by a motion axis and the extents of the motion.

Example:
\verbatim
    dynamicFvMesh   dynamicMotionSolverFvMesh;

    motionSolver    displacementLinearMotion;

    axis            (0 1 0);
    xFixed          0.8;
    xMoving         0;
    displacement    table
    (
        (0 0)
        (4 0.7)
    );
\endverbatim

This mesh is compressed between \c xFixed and \c xMoving in the direction
\c axis between time 0 and 4 with a maximum displacement of 0.7.
The mesh beyond \c xFixed is fixed and beyond \c xMoving moves with maximum
displacement.